Added support for spellchecking multilanguage documents
The language is determined contextually by the lang
attribute. Because the language of a Text::Aspell object cannot
be changed dynamically, we maintain spellcheckers for each
language that is requested.

Added a langs method for retrieving a list of languages that
there are spellcheckers for.

Fixed bug with encoding the text.

Updated documentation accordingly.

TODO - Add a means of restricting the languages checked, and
verifying that a dictionary for that language is installed.